Support a dep cache for build, build-complete, and build-snapshot
This feature takes the openssl, lzo, and pkcs11-helper dependency builds
and creates a "depcache" tarball for each dep making it available for
use in future builds without recompiling them.

build-complete has a high-level feature that can build or use a depcache
with further docs in the updated README. This feature is available to
build-snapshot only for use (create the depcache with build-complete.)

For the generic depcache support, a named depcache can be passed on the
command-line to save or use the depcache for that run.

Advanced use via env-vars is also supported for environments that wish
to use it via the SAVE_DEP_CACHE, USE_DEP_CACHE, and DEPCACHE_SUFFIX;
note that this applies to the generic/build script only.

Signed-off-by: Josh Cepek <josh.cepek@usa.net>