end running openvpn-gui.exe process
The installer now checks if openvpn-gui.exe is running. If it is, the user must agree to close the GUI for the install to continue. A new checkbox on the finish page "Start OpenVPN GUI" is automatically checked if the GUI was killed, it is left unchecked if not. The uninstaller will stop openvpn-gui.exe without prompting. 
String clean up.
Switch to NSIS MUI2 (vs MUI).