Disable broken "make check" on jessie and trusty
It is not clear why these two operating systems now run "make check" by default,
because this was not the case when OpenVPN 2.3.9 was built a few weeks ago. In
any case, the tests were now forcibly disabled to allow building packages for
these platforms. The actual test failure in t_cltsrv.sh was not surprising,
given that the test was running inside a schroot.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>