windows-msi: Move license agreement to external file
The license agreement page of MSI GUI is not displayed by default. It
accidentially contained GPLv3 license RTF as a one huge XML attribute value.
Not only was the license wrong, keeping an RTF text as an XML attribute
value does not scale (line endings, XML escaping, etc.)

This patch moves the license agreement into an external RTF file while the
GPLv3 text was replaced by a stub text.

Should we ever require to display the license agreement to end-user before
installing, we are ready for it.

Signed-off-by: Simon Rozman <simon@rozman.si>