windows-msi: Arrange services and OpenVPN GUI to stop/start on update
In order to start using the new version of OpenVPN after an update the
OpenVPN GUI and OpenVPN services must be recycled.

This patch stops all services on install too. This terminates not only the
services but all `openvpn.exe` processes managed by the interactive service
as well.

Furthermore, in the full UI MSI version the OpenVPN GUI is recycled as well.
Please note that stopping the OpenVPN GUI works for non-elevated
`openvpn-gui.exe` processes only. Older versions before the interactive
service was invented and run elevated are not stopped.

When user clicks the "Close" button after a successful installation, the
OpenVPN GUI is launched back. If the OpenVPN GUI is chosen to launch on
startup of course.

Signed-off-by: Simon Rozman <simon@rozman.si>