windows-msi: Disable openvpnserv2.exe feature when .NETFx4 is not installed
The `openvpnserv2.exe` requires .NET Framework 4.0 Full. This patch sets the
install level for "OpenVPN.Service" feature to 0 if the .NET Framework is
not detected. This effectively disables (and hides) this feature preventing
its installation.

Should we later decide to start the `openvpnserv2.exe` and set its startup
to Automatic in the setup process, this check&disable will prevent setup
failures on computers without .NET Framework 4.0. Note that a failure to
start a service is fatal for MSI setup process.

Signed-off-by: Simon Rozman <simon@rozman.si>