windows-msi: Switch to public version of DriverCertification property
The FindSystemInfo custom action sets DRIVERCERTIFICATION (public) instead
of DriverCertification (private) now.  This patch adopts to this change, and
modifies FindSystemInfo authoring to make MSI call it only in the first
sequence found.

BTW, the description of FindSystemInfo was made more generic, as this action
will soon be extended with additional work to do.

Signed-off-by: Simon Rozman <simon@rozman.si>