windows-msi: Demote openvpnserv2.exe feature and enable it if used
The openvpnserv2.exe service will not install by default any more.  It is an
opt-in feature now.  However, if the MSI detects the service was running or
was configured for auto-start when updating, it will select the feature for
install.  This will auto-enable the openvpnserv2.exe service for users
actually using it when upgrading NSIS installation with an MSI.  Mind that
feature selection is preserved on MSI>>MSI updates.

Signed-off-by: Simon Rozman <simon@rozman.si>