windows-msi: Ignore StubPath registry value delete failures on uninstall
WScript.Shell's method RegDelete is silly enough to throw an error when
a registry value it is trying to delete something that does not exist.
This is not really an error condition as the state we are trying to
achieve is already achieved.

In case there is no StubPath value in the registry, PublishActiveSetup()
fails in the uninstall phase preventing OpenVPN uninstallation.

Signed-off-by: Simon Rozman <simon@rozman.si>