windows-msi: Fix OpenVPNService upgrading
When MSI detects user is using OpenVPNService (the service is running or set
to auto-start), it marks OpenVPN.Service for install. If the config-auto
subfolder does not exist yet, it moves all files from config subfolder to
config-auto. This is a one-time migration of auto-start config files.

Another feature of the MSI installer is a thorough cleanup of all the NSIS
files and registry settings before install.

So far so good. But (there always is a "but", otherwise you wouldn't be
reading this, would you?), both features combined explode!

The RemoveFiles removes the NSIS' config/README.txt file first, while the
MoveFiles action already has a list of files to move from config to
config-auto. With config/README.txt already gone, MoveFiles fails.

I have decided not to clean any NSIS leftover files from the config
subfolder. The README.txt file is overwriten by the same-named MSI supplied
file anyway. And the config subfolder is kind of sacred => it contains user
authored files. Moving them to a different folder is bold enough.

Signed-off-by: Simon Rozman <simon@rozman.si>