windows-msi: Disable change and repair in MSI GUI
This commits inhibits Change and Repair in the MSI GUI, as they do not
work correctly with Wintun and TAP-Windows6 setup.

Rather than completely removing the GUI options, they were merely
disabled, as the Wintun and TAP-Windows6 driver setup will be rewritten
in the near future to allow proper MSI operation. Then, the Change and
Repair will be available again.

Signed-off-by: Simon Rozman <simon@rozman.si>