msvc: Normalize downloaded filenames
When using MSVC build on GitHub branches the URL filename is not a
proper "<project>-<version>.tar.gz", but rather a "master.tar.gz"
confusing the build script.

The build script was modified to use exact filenames in sources folder:
<openssl|lzo|pkcs11|tap>-<version>[.tar].<extension>

All dependencies have _VERSION variable defined in build-env.bat.
Therefore, this information is used to provide <version> in the filename
above. Just keep in mind when version is "master", you need to clean the
cache in sources folder when a newer version of "master" is to be built.

Project folder names in build.tmp are still governed by the package
TAR/Zip archive.

Signed-off-by: Simon Rozman <simon@rozman.si>