generic/build: build lz4 as dynamic library
To be consistent with other dependencies and to avoid
cludge in build system to support both mingw and msvc
openvpn binaries, build lz4 as a dynamic library.

To make this change ARM64-proof, MSI part includes
customization of lz4 dll name. MinGW builds it as "liblz4.dll",
but MSVC (will be used for ARM64) uses "lz4.dll".

Signed-off-by: Lev Stipakov <lev@openvpn.net>