windows-msi: add vcruntime.dll for arm64
ARM64 binaries are built with MSVC, therefore MSVC library
needs to be bundled. The recommended way to include it is to use MSMs:

https://devblogs.microsoft.com/cppblog/how-to-redistribute-the-visual-c-libraries-with-your-application/

However as for now there is no MSM for ARM64 available, so we do
"app-local" deployment and copy required DLL to bin directory.

Since as for now we only build ARM64 binaries with MSVC, no need
to provide that DLL for another (x86, x64) architectures.

Signed-off-by: Lev Stipakov <lev@openvpn.net>