OpenVPNService: install by default and keep previous state
As we don't support "Modify" action, adding OpenVPNService
is problematic since it requires uninstall and reinstall.

Install it by default. For clean install, use "Manual"
start mode and doesn't start it. For upgrade, preserve existing
configuration - for example if start mode was "automatic" and service
was started - the same will apply for the new installation.

This allows to remove code from openvpnmsica.dll which
reads service status, but doesn't apply it to the new installation.

Signed-off-by: Lev Stipakov <lev@openvpn.net>