windows-msi/version.m4: Update Easy-RSA v3.1.0 to v3.1.1
Update download source to: OpenVPN/easy-rsa Release v3.1.1

Add comments to clarify that the 64bit version of Easy-RSA is suitable
for OpenVPN Windows Installer for both 32bit and 64bit.

For the record:
The only part of Easy-RSA which is 64bit is the pre-built OpenSSL binary.
The Easy-RSA for Windows zip files, distributed from github, do not contain
any OpenSSL binaries.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>