src: add submodules for every git repository
This way we can cleanly describe with a single tag
in this repository what we used for building. Of course
this does not cover every build dependency, but the
previous way of just doing "git pull" in build-and-package.ps1
was very unreliable.

To do a build against "latest" git submodule update --remote
can be used.

Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>