Fix DCO adapter removal when OpenVPN Connect is installed
At uninstall, we run EvaluateTUNTAPAdapters CA to determine
if we should add or remove an adapter. This is done by checking
requested action of component.

When upcoming version of OpenVPN Connect is installed alongside with
OpenVPN GUI and then GUI is uninstalled, MSI disallows component
(CMP_ovpn_dco.inf) uninstall, since it has been used by Connect,
and sets requested action to INSTALLSTATE_UNKNOWN. In this case we don't
remove an adapter.

Fix by changing component to "shortcut.bin.tapctl.exe.create.dco" - it is
tied to DCO feature, but openvpn2-specific and is installed/removed
along with openvpn2.

Signed-off-by: Lev Stipakov <lev@openvpn.net>