GHA: do not use latest vcpkg, use the exact submodule
The MSI installer depends on openssl.exe. That is now
provided by the "tools" feature of the openssl vcpkg
package. But it is unclear where to add that dependency.
Neither openvpn nor openvpn-gui need the openssl.exe
for building. Only the MSI depends on it, but that doesn't
have its own vcpkg build.

It was moved to a separate feature to avoid incurring a
pwsh dependency for Linux minGW builds. So we might not
want to add that dependency back unnecessarily.

Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>