windows-msi: don't depend on core.autocrlf=false
The build script assumed that the sample files in
openvpn were checked out with LF line-endings. Since
these files have no eol enforced by .gitattributes in
openvpn, this depends on core.autocrlf=false, which
is not the default on Windows.

So assume that these files will be checked out with
CRLF line endings on Windows. The alternative would
be to enforce LF line-endings via openvpn's
.gitattributes.

Fixes #358

Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>