GHA: check out the actual state of the submodules
Previously we checked out the current remote state of
the submodule branch. This was required for the builds
triggered remotely by openvpn changes.

Now that we have renovate to trigger builds on submodule
remote changes we do not need this anymore. And it actually
hurts since the PR builds do not actually test the changes
in the PR in some cases.

Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>